Switched much of the code to use genotype categories for inference, instead of keeping with c_gt, v_gt, etc.
I also added code to read in a table of simulations, which has to include n_der and n_tot, and to construct a sims.dat from that.
This removes the ability okkkkhuhyuuuuyyyyyyyyyyym,,. [edit by hugo.]

Continuing where I left off - this removes thee abilityto do the "full" EM model, where p(der) at each point is estimated by
fitting a curve to many simulations.  I did some testing for this, and for the situations that I examined [various points on
the vindija lineage] it made only a slight difference to use the simple vs full model.

Many things in this commit still do not use gt.category, but it does run the EM successfully.